Move CombinedTree to incrementalmerkletree::testing
